Refactor Logica Repetida en tabla de oradores.
Creacion del componente TablaOradores que reemplaza ListaPinosQueHablaron.js
Separada en otro archivo la logica de la fila de esa tabla.
